+++ client-v2/src/main/java/com/clickhouse/client/api/InsertResponse.java
@@ -1,0 +1,4 @@
+package com.clickhouse.client.api;
+
+public class InsertResponse {
+}