+++ client-v2/src/main/java/com/clickhouse/client/api/TableSchema.java
@@ -1,0 +1,5 @@
+package com.clickhouse.client.api;
+
+public class TableSchema {
+
+}